requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.0.4
django-admin-rangefilter==0.12.4
django-allauth==0.62.1
django-cors-headers==4.3.1
django-youtube==0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-api-keys==0.5.9
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
msgpack==1.0.8
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.20.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uWSGI==2.0.25.1
webdriver-manager==4.0.1
whitenoise==6.6.0
wsproto==1.2.0
